uji: clunker together a test description section
The test description is added as a paragraph with a h3 beforehand.
The tests from that section are collected too into the higher-level section,
though the actual description can't be fetched through uji test yet.

This works for now to tick the tests off, but really we need to figure out the
markdown structure we'll stick with and then adjust the parsing accordingly.