uji new: append .yaml if necessary
For invocations of `uji new foo`, try if `foo.yaml` or `foo.yml` exists.
If so, use that as template.

Once we have a foo.yaml and foo-1234/ test directory, shell completion
may stop after "foo", so auto-completing the yaml bit makes `uji new` a
bit nicer to use.

Fixes #7