uji view: support rendering a help screen
Press ? to get a list of key mappings.

A larger patch than anticipated because it centralizes the keymap
handling in one spot via a custom class + flags. Those flags are then
used to check if/where we need special handlings - e.g. during the help
screen the movements are disabled.

Fixes #10