uji view: switch from curtsies to blessed
blessed is a lot simpler than curtsies, and it supports truecolor which
allows us to use better styles in the future.

blessed works with simple printfs, so our rendering process is now just
that - a bunch of printfs followed by one for the status line, followed
by the cursor icon.