uji view: make the rendering loop less confusing
render/redraw wasn't named usefully - switch to a rerender() to render
to the line buffer and _display() to actually display on screen.

And add a dirty property to trigger rerender automatically - this way we
have only one location we need to do this. Except for execute_command
which needs to do this live though.